feat : 로또(자동)
추가내용
* 당첨정보객체에서 가지고 있는 당첨정보와 주입받은 구매로또 정보들을 이용해서 로또당첨결과인 LottoResult를 반환 하도록 기존 서비스영역에 코드를 이관 후 수정함.
* 기존 당첨매칭까지 수행하던 LotteryDraw 서비스는 구매로또정보를 주입 받을 필요가 없어지고, 로또당첨결과내용을 받아서 수익률 계산을 담당하도록 역할이 축소됨.
* 기존에 lotto의 getter을 통해 접근하여 담겨있는 로또 번호 확인하는 방식을 lotto 객체에 메세지를 통해서 결과를 받는 방식으로 수정함.
