+++ src/main/java/lotto/domain/WinLottoInfo.java
@@ -2,16 +2,21 @@ package lotto.domain;
 
 import java.util.List;
 import lotto.message.Message;
+import lotto.service.Rank;
 
 public class WinLottoInfo {
 
-  private Lotto lotto;
+  private static final int INT_ZERO = 0;
+
+  private Lotto winLotto;
 
   private LottoNumber bonusNumber;
 
+  private LottoResult lottoResult;
+
   public WinLottoInfo(final Lotto lotto, final int bonus) {
     validation(lotto, bonus);
-    this.lotto = lotto;
+    this.winLotto = lotto;
     this.bonusNumber = new LottoNumber(bonus);
   }
 
@@ -21,8 +26,8 @@ public class WinLottoInfo {
     }
   }
 
-  public Lotto getLotto() {
-    return lotto;
+  public Lotto getWinLotto() {
+    return winLotto;
   }
 
   public LottoNumber getBonusLottoNumber() {
@@ -30,6 +35,49 @@ public class WinLottoInfo {
   }
 
   public List<LottoNumber> getLottoNumbers(){
-    return lotto.getLotto();
+    return winLotto.getLotto();
+  }
+
+  public LottoResult matchLottoInfo(Lotteries lotteries) {
+    return getInputMatchTotalInfo(lotteries);
+  }
+
+  public LottoResult getInputMatchTotalInfo(final Lotteries lotteries) {
+    lottoResult = new LottoResult();
+
+    for (Lotto lotto : lotteries.getLottos()) {
+      getLottosByRank(lotto).add(lotto);
+    }
+
+    lottoResult.getCategoriesRank().remove(Rank.MISS);
+    return lottoResult;
+  }
+
+  private List<Lotto> getLottosByRank(final Lotto lotto) {
+    return lottoResult.getMatchLottos(getMatchCountForRank(lotto));
+  }
+
+  private Rank getMatchCountForRank(Lotto lotto) {
+    return Rank.matchRank(getCountByRank(lotto, INT_ZERO),
+        isMatchBonus(lotto));
+  }
+
+  private int getCountByRank(final Lotto lotto, int count) {
+    for (LottoNumber lottoNumber : winLotto.getLotto()) {
+      count = checkContainValues(lotto, count, lottoNumber);
+    }
+
+    return count;
+  }
+
+  private int checkContainValues(final Lotto lotto, int count, final LottoNumber lottoNumber) {
+    if (lotto.match(lottoNumber)) {
+      count++;
+    }
+    return count;
+  }
+
+  private boolean isMatchBonus(final Lotto lotto) {
+    return lotto.getLotto().contains(bonusNumber);
   }
 }
