+++ src/test/java/lotto/service/LotteryDrawTest.java
@@ -1,15 +1,10 @@
 package lotto.service;
 
-import static java.util.stream.IntStream.range;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
-import java.util.List;
 import lotto.domain.Lotteries;
 import lotto.domain.Lotto;
 import lotto.domain.LottoMoney;
-import lotto.domain.LottoResult;
-import lotto.domain.WinLottoInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -19,57 +14,17 @@ class LotteryDrawTest {
   @Test
   void 당첨번호로또생성() {
 
-    LotteryDraw lotteryDraw = new LotteryDraw(new Lotteries(), new LottoMoney(1000));
+    LotteryDraw lotteryDraw = new LotteryDraw(new LottoMoney(1000));
     Lotto lotto = lotteryDraw.inputWinningNumbers("1,2,3,4,5,6");
 
     assertThat(lotto.getLotto().size()).isEqualTo(6);
   }
 
-  @DisplayName("당첨번호와 로또 비교하여 해당 등수에 로또객체를 적재하는 테스트.")
-  @Test
-  void 당첨번호와로또비교후등수별정리() {
-
-    List<Integer> values = new ArrayList<>();
-    range(1, 7).forEach(values::add);
-    Lotto testLotto = new Lotto(values);
-
-    Lotteries lotteries = new Lotteries();
-    List<Lotto> testLottos = lotteries.getLottos();
-    testLottos.add(testLotto);
-
-    LotteryDraw lotteryDraw = new LotteryDraw(lotteries, new LottoMoney(1000));
-
-    WinLottoInfo winLottoInfo = lotteryDraw.createWinLottoInfo("4,5,6,11,12,13", 21);
-    LottoResult lottoResult = lotteryDraw.matchLottoInfo(winLottoInfo);
-
-    assertThat(lottoResult.getCategoriesRank().get(Rank.FIFTH).size()).isEqualTo(1);
-  }
-
   @DisplayName("당첨로또와 보너스번호를 가진 당첨로또정보객체 생성")
   @Test
   void 당첨로또정보객체생성() {
-    LotteryDraw lotteryDraw = new LotteryDraw(new Lotteries(), new LottoMoney(1000));
+    LotteryDraw lotteryDraw = new LotteryDraw(new LottoMoney(1000));
     lotteryDraw.createWinLottoInfo("1,2,3,4,5,7",6);
   }
 
-  @DisplayName("당첨번호,보너스번호를  로또 비교하여 해당 등수에 로또객체를 적재하는 테스트.")
-  @Test
-  void 당첨번호와보너스번호를로또비교후등수별정리() {
-
-    List<Integer> values = new ArrayList<>();
-    range(1, 7).forEach(values::add);
-    Lotto testLotto = new Lotto(values);
-
-    Lotteries lotteries = new Lotteries();
-    List<Lotto> testLottos = lotteries.getLottos();
-    testLottos.add(testLotto);
-
-    LotteryDraw lotteryDraw = new LotteryDraw(lotteries, new LottoMoney(1000));
-
-    WinLottoInfo winLottoInfo = lotteryDraw.createWinLottoInfo("1,2,3,4,5,11", 6);
-
-    LottoResult lottoResult = lotteryDraw.matchLottoInfo(winLottoInfo);
-
-    assertThat(lottoResult.getCategoriesRank().get(Rank.SECOND).size()).isEqualTo(1);
-  }
 }
